+++ DomainServices/Interfaces/ICustomerServices.cs
@@ -7,7 +7,7 @@ public interface ICustomerServices
 {
     IEnumerable<Customer> GetAll();
     Customer GetById(long id);
-    int Create(Customer model);
+    long Create(Customer model);
     bool Update(Customer model);
     bool Delete(long id);
 }
